relocate conversion traits and fix build
